+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The types of changes are:
 * Add authentication to the `masking` endpoints as well as accompanying scopes [#2909](https://github.com/ethyca/fides/pull/2909)
 * Add an Organization Management page (beta) [#2908](https://github.com/ethyca/fides/pull/2908)
 * Adds assigned systems to user management table [#2922](https://github.com/ethyca/fides/pull/2922)
+* APIs to support Privacy Notice management (create, read, update) [#2928](https://github.com/ethyca/fides/pull/2928)
 
 ### Changed